fix(FitHelper): widen byte-offset locals to uint16_t for >85-field defs
Two byte-offset locals were sized for the bundled EXAMPLE-profile FIT
SDK and silently wrap on RELEASE-profile defs:

1. makeMsgDef() copies optimised field defs out of the origin def at
   byte offset 3*idx.  AVG_HEART_RATE sits at idx 88 in the release
   lap def, so 3*88 = 264 truncates to 8 in uint8_t and the wrong
   bytes get copied -- producing the (133, 4, 4) / (133, 5, 4)
   triples that fitfileviewer.com flags as "Duplicate field
   definition: field_133" / "...avg_step_length" in lap and session.

2. getFieldOffset() accumulates per-field byte offsets across the C
   message struct.  FIT_SESSION_MESG is 340 bytes in the release
   profile, so the uint8_t accumulator wraps for late-in-struct
   fields and the writer ends up reading neighbouring bytes for
   total_calories / metabolic_calories -- which is why a small
   indoor session was reporting tens of thousands of kcal.

Both bugs were dormant on the EXAMPLE profile because lap/session
defs had <=95 fields and structs <=234 bytes, so 3*idx and the
running offset always fit in uint8_t.

Widen:
- makeMsgDef src/dst byte-offset locals to uint16_t.
- getFieldOffset return type and accumulator to uint16_t (UINT16_MAX
  sentinel).
- MsgField::offset/size to uint16_t (used to write field data, with
  full-message structs >255 bytes a possibility on newer profiles).
- makeMsgFields full-message size accumulator to uint16_t.

This fix benefits any current or future FitHelper user that links
against a larger profile; existing apps on the EXAMPLE profile are
unaffected because they never exercised the overflow paths.

+++ Libs/Source/FitHelper/FitHelper.cpp
@@ -283,13 +283,17 @@ namespace SDK::Component {
         for (auto f : fields) {
             for (uint8_t idx = 0; idx < mMsgDefOrigin->num_fields; ++idx) {
                 if (f == mMsgDefOrigin->fields[FIT_MESG_DEF_FIELD_OFFSET(field_def_num, idx)]) {
-                    uint8_t field_def_num_dst = FIT_MESG_DEF_FIELD_OFFSET(field_def_num, offset);
-                    uint8_t size_dst          = FIT_MESG_DEF_FIELD_OFFSET(size, offset);
-                    uint8_t base_type_dst     = FIT_MESG_DEF_FIELD_OFFSET(base_type, offset);
+                    // 16-bit byte offsets into the def buffer.  Profiles like
+                    // RELEASE have lap/session defs with >85 fields, so 3*idx
+                    // overflows uint8_t for high-index fields (e.g. avg_heart_rate
+                    // sits at idx 88 in the release lap def -> 264 bytes).
+                    uint16_t field_def_num_dst = FIT_MESG_DEF_FIELD_OFFSET(field_def_num, offset);
+                    uint16_t size_dst          = FIT_MESG_DEF_FIELD_OFFSET(size, offset);
+                    uint16_t base_type_dst     = FIT_MESG_DEF_FIELD_OFFSET(base_type, offset);
 
-                    uint8_t field_def_num_src = FIT_MESG_DEF_FIELD_OFFSET(field_def_num, idx);
-                    uint8_t size_src          = FIT_MESG_DEF_FIELD_OFFSET(size, idx);
-                    uint8_t base_type_src     = FIT_MESG_DEF_FIELD_OFFSET(base_type, idx);
+                    uint16_t field_def_num_src = FIT_MESG_DEF_FIELD_OFFSET(field_def_num, idx);
+                    uint16_t size_src          = FIT_MESG_DEF_FIELD_OFFSET(size, idx);
+                    uint16_t base_type_src     = FIT_MESG_DEF_FIELD_OFFSET(base_type, idx);
 
                     writable->fields[field_def_num_dst] = mMsgDefOrigin->fields[field_def_num_src];
                     writable->fields[size_dst]          = mMsgDefOrigin->fields[size_src];
@@ -340,7 +344,9 @@ namespace SDK::Component {
         if (fields.size() == 0) {
 		    mMsgFields.clear();
 
-            uint8_t size = 0;
+            // 16-bit accumulator: full message structs in newer profiles can
+            // exceed 255 bytes (e.g. FIT_SESSION_MESG = 340 bytes).
+            uint16_t size = 0;
             for (uint8_t idx = 0; idx < mMsgDefOrigin->num_fields; ++idx) {
                 size += mMsgDefOrigin->fields[FIT_MESG_DEF_FIELD_OFFSET(size, idx)];
             }
@@ -362,9 +368,13 @@ namespace SDK::Component {
         }
     }
 
-    uint8_t FitHelper::getFieldOffset(FIT_EVENT_FIELD_NUM field) const
+    uint16_t FitHelper::getFieldOffset(FIT_EVENT_FIELD_NUM field) const
     {
-        uint8_t offset = 0;
+        // 16-bit accumulator -- C message structs in newer profiles can be
+        // >255 bytes (e.g. FIT_SESSION_MESG = 340 bytes), so summing per-field
+        // sizes overflows uint8_t and produces wildly wrong byte offsets when
+        // the writer later reads field data out of the caller's struct.
+        uint16_t offset = 0;
         for (FIT_UINT8 idx = 0; idx < mMsgDefOrigin->num_fields; ++idx) {
             if (field == mMsgDefOrigin->fields[FIT_MESG_DEF_FIELD_OFFSET(field_def_num, idx)]) {
                 return offset;
@@ -372,14 +382,14 @@ namespace SDK::Component {
         
             FIT_UINT8 base_type_num = mMsgDefOrigin->fields[FIT_MESG_DEF_FIELD_OFFSET(base_type, idx)] & FIT_BASE_TYPE_NUM_MASK;
             if (base_type_num >= FIT_BASE_TYPES) {
-                return UINT8_MAX;
+                return UINT16_MAX;
             }
 
             FIT_UINT8 fieldSize = mMsgDefOrigin->fields[FIT_MESG_DEF_FIELD_OFFSET(size, idx)];
             offset += fieldSize;
         }
 	
-        return UINT8_MAX;
+        return UINT16_MAX;
     }
 
     uint8_t FitHelper::getFieldSize(FIT_EVENT_FIELD_NUM field) const
